Reports are almost done

+++ src/main/java/com/example/PAPS/dtos/ClientDto.java
@@ -11,6 +11,6 @@ public class ClientDto {
 
     private Long id;
     private String FIO;
-    private Integer passportID;
+    private Long passportID;
     private String address;
 }
